commented out for the earch-functionality branch to save our API requests; added TableDisplay.js reworked tmGetEventsButton event listener to now use new displayConcertSearchResults from TableDisplay.js
