+++ app/public/search.js
@@ -1,10 +1,11 @@
+
 let submitSearchButton = document.getElementById("submitSearchButton");
 let artistInput = document.getElementById("artist");
 let genreInput = document.getElementById("genre");
 let locationInput = document.getElementById("location");
 
 function submitSearch(){
-    console.log("asdf");
+    console.log("Meyo");
 }
 
 submitSearchButton.addEventListener("click", submitSearch);
